complete

+++ W26W10WeatherApp/WeatherProxy.cs
@@ -31,14 +31,14 @@ namespace W26W10WeatherApp
 
     public class Main
     {
-        public int temp { get; set; }
+        public double temp { get; set; }
         public double feels_like { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
-        public int sea_level { get; set; }
-        public int grnd_level { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
+        public double sea_level { get; set; }
+        public double grnd_level { get; set; }
     }
 
     public class Rain
